+++ arviz/plots/backends/matplotlib/bpvplot.py
@@ -13,7 +13,7 @@ from ...plot_utils import (
     is_valid_quantile,
     matplotlib_kwarg_dealiaser,
 )
-from ....kde_utils import kde
+from ....kde_utils import _kde
 
 def plot_bpv(
     ax,
@@ -76,7 +76,7 @@ def plot_bpv(
 
         if kind == "p_value":
             tstat_pit = np.mean(pp_vals <= obs_vals, axis=-1)
-            x_s, tstat_pit_dens = kde(tstat_pit)
+            x_s, tstat_pit_dens = _kde(tstat_pit)
             ax_i.plot(x_s, tstat_pit_dens, linewidth=linewidth, color=color)
             ax_i.set_yticks([])
             if reference is not None:
@@ -95,7 +95,7 @@ def plot_bpv(
 
         elif kind == "u_value":
             tstat_pit = np.mean(pp_vals <= obs_vals, axis=0)
-            x_s, tstat_pit_dens = kde(tstat_pit)
+            x_s, tstat_pit_dens = _kde(tstat_pit)
             ax_i.plot(x_s, tstat_pit_dens, color=color)
             if reference is not None:
                 if reference == "analytical":