+++ arviz/plots/backends/matplotlib/forestplot.py
@@ -10,7 +10,7 @@ from . import backend_kwarg_defaults, backend_show
 from ....stats import hdi
 from ....stats.diagnostics import _ess, _rhat
 from ....numeric_utils import histogram, get_bins
-from ....kde_utils import kde
+from ....kde_utils import _kde
 from ...plot_utils import _scale_fig_size, xarray_var_iter, make_label
 from ....utils import conditional_jit
 
@@ -519,7 +519,7 @@ class VarHandler:
                 density_q = density.cumsum() / density.sum()
                 x = x[:-1]
             elif kind == "density":
-                x, density = kde(values)
+                x, density = _kde(values)
                 density_q = density.cumsum() / density.sum()
 
             xvals.append(x)