+++ arviz/plots/backends/matplotlib/ppcplot.py
@@ -13,7 +13,7 @@ from ...plot_utils import (
     _scale_fig_size,
 )
 from ....numeric_utils import histogram, get_bins
-from ....kde_utils import kde
+from ....kde_utils import _kde
 _log = logging.getLogger(__name__)
 
 
@@ -149,7 +149,7 @@ def plot_ppc(
             for vals in pp_sampled_vals:
                 vals = np.array([vals]).flatten()
                 if dtype == "f":
-                    pp_x, pp_density = kde(vals)
+                    pp_x, pp_density = _kde(vals)
                     pp_densities.append(pp_density)
                     pp_xs.append(pp_x)
                 else:
@@ -367,7 +367,7 @@ def _set_animation(
     if kind == "kde":
         length = len(pp_sampled_vals)
         if dtype == "f":
-            x_vals, y_vals = kde(pp_sampled_vals[0])
+            x_vals, y_vals = _kde(pp_sampled_vals[0])
             max_max = max([max(kde(pp_sampled_vals[i])[1]) for i in range(length)])
             ax.set_ylim(0, max_max)
             (line,) = ax.plot(x_vals, y_vals, **plot_kwargs)