+++ arviz/plots/backends/matplotlib/violinplot.py
@@ -4,7 +4,7 @@ import numpy as np
 
 from . import backend_show
 from ....stats import hdi
-from ....kde_utils import kde
+from ....kde_utils import _kde
 from ....numeric_utils import histogram, get_bins
 from ...plot_utils import make_label, _create_axes_grid, matplotlib_kwarg_dealiaser, _scale_fig_size
 
@@ -23,6 +23,7 @@ def plot_violin(
     rug_kwargs,
     bw,
     textsize,
+    circular,
     hdi_prob,
     quartiles,
     backend_kwargs,
@@ -60,7 +61,7 @@ def plot_violin(
         if val[0].dtype.kind == "i":
             dens = cat_hist(val, rug, shade, ax_, **shade_kwargs)
         else:
-            dens = _violinplot(val, rug, shade, bw, ax_, **shade_kwargs)
+            dens = _violinplot(val, rug, shade, bw, circular, ax_, **shade_kwargs)
 
         if rug:
             rug_x = -np.abs(np.random.normal(scale=max(dens) / 3.5, size=len(val)))
@@ -85,9 +86,14 @@ def plot_violin(
     return ax
 
 
-def _violinplot(val, rug, shade, bw, ax, **shade_kwargs):
+def _violinplot(val, rug, shade, bw, circular, ax, **shade_kwargs):
     """Auxiliary function to plot violinplots."""
-    x, density = kde(val, bw_fct=bw)
+    if bw is "default":
+        if circular:
+            bw = "taylor"
+        else:
+            bw = "experimental"
+    x, density = _kde(val, circular=circular, bw=bw)
 
     if not rug:
         x = np.concatenate([x, x[::-1]])