+++ arviz/plots/kdeplot.py
@@ -4,7 +4,7 @@ import xarray as xr
 
 from ..data import InferenceData
 from ..numeric_utils import _fast_kde_2d
-from ..kde_utils import kde
+from ..kde_utils import _kde
 from .plot_utils import get_plotting_function
 from ..rcparams import rcParams
 
@@ -15,7 +15,9 @@ def plot_kde(
     cumulative=False,
     rug=False,
     label=None,
-    bw=1,
+    bw="default",
+    adaptive=False,
+    circular=False,
     quantiles=None,
     rotated=False,
     contour=True,
@@ -51,10 +53,19 @@ def plot_kde(
         If True adds a rugplot. Defaults to False. Ignored for 2D KDE
     label : string
         Text to include as part of the legend
-    bw: float, optional
-        Bandwidth scaling factor for 1D KDE. Must be larger than 0.
-        The higher this number the smoother the KDE will be.
-        Defaults to 1 which means the bandwidth is not modified.
+    bw: float or str, optional
+        If numeric, indicates the bandwidth and must be positive.
+        If str, indicates the method to estimate the bandwidth and must be
+        one of "scott", "silverman", "isj" or "experimental" when `circular` is False
+        and "taylor" (for now) when `circular` is True.
+        Defaults to "default" which means "experimental" when variable is not circular 
+        and "taylor" when it is.
+    adaptive: bool, optional. 
+        If True, an adaptative bandwidth is used. Only valid for 1D KDE.
+        Defaults to False.
+    circular: bool, optional.
+        If True, it interprets `values` is a circular variable measured in radians
+        and a circular KDE is used. Only valid for 1D KDE. Defaults to False.
     quantiles : list
         Quantiles in ascending order used to segment the KDE.
         Use [.25, .5, .75] for quartiles. Defaults to None.
@@ -129,6 +140,35 @@ def plot_kde(
 
         >>> az.plot_kde(mu_posterior, rug=True)
 
+    Plot KDE with adaptive bandwidth
+
+    .. plot::
+        :context: close-figs
+
+        >>> az.plot_kde(mu_posterior, adaptive=True)
+
+    Plot KDE with a different bandwidth estimator
+
+    .. plot::
+        :context: close-figs
+
+        >>> az.plot_kde(mu_posterior, bw="scott")
+
+    Plot KDE with a bandwidth specified manually
+
+    .. plot::
+        :context: close-figs
+
+        >>> az.plot_kde(mu_posterior, bw=0.4)
+
+    Plot KDE for a circular variable
+
+    .. plot::
+        :context: close-figs
+
+        >>> rvs = np.random.vonmises(mu=np.pi, kappa=2, size=500)
+        >>> az.plot_kde(rvs, circular=True)
+
 
     Plot a cumulative distribution
 
@@ -180,6 +220,10 @@ def plot_kde(
         :context: close-figs
 
         >>> az.plot_kde(mu_posterior, values2=tau_posterior, contour=False)
+        
+    See Also
+    --------
+    plot_kde : Compute and plot a kernel density estimate.
 
     """
     if isinstance(values, xr.Dataset):
@@ -194,7 +238,14 @@ def plot_kde(
         )
 
     if values2 is None:
-        grid, density = kde(values, bw_fct=bw, cumulative=cumulative)
+        
+        if bw is "default":
+            if circular:
+                bw = "taylor"
+            else:
+                bw = "experimental"
+
+        grid, density = _kde(values, circular, bw=bw, adaptive=adaptive, cumulative=cumulative)
         lower, upper = grid[0], grid[-1]
 
         if cumulative: