+++ arviz/tests/base_tests/test_utils.py
@@ -5,6 +5,7 @@ Tests for arviz.utils.
 from unittest.mock import Mock
 import numpy as np
 import pytest
+import scipy.stats as st
 
 from arviz.data.base import dict_to_dataset
 from ...utils import (
@@ -17,6 +18,7 @@ from ...utils import (
     _subset_list,
 )
 from ...data import load_arviz_data, from_dict
+from ...numeric_utils import _circular_mean, _normalize_angle
 
 
 @pytest.fixture(scope="session")
@@ -271,3 +273,23 @@ def test_flatten_inference_data_to_dict(
                 assert not any("sample_stats" in item for item in res_dict)
         else:
             assert not any("prior" in item for item in res_dict)
+
+
+@pytest.mark.parametrize("mean", [0, np.pi, 4 * np.pi, - 2 * np.pi, -10 * np.pi])
+def test_circular_mean_scipy(mean):
+    """Test our `_circular_mean()` function gives same result than Scipy version."""
+    rvs = st.vonmises.rvs(loc=mean, kappa=1, size=1000)
+    mean_az = _circular_mean(rvs)
+    mean_sp = st.circmean(rvs, low=-np.pi, high=np.pi)
+    np.testing.assert_almost_equal(mean_az, mean_sp)
+
+
+@pytest.mark.parametrize("mean", [0, np.pi, 4 * np.pi, - 2 * np.pi, -10 * np.pi])
+def test_normalize_angle(mean):
+    """Testing _normalize_angles() return values between expected bounds"""
+    rvs = st.vonmises.rvs(loc=mean, kappa=1, size=1000)
+    values = _normalize_angle(rvs, zero_centered=True)
+    assert ((-np.pi <= values) & (values <= np.pi)).all()
+
+    values = _normalize_angle(rvs, zero_centered=False)
+    assert ((0 <= values) & (values <= 2 * np.pi)).all()